+++ framework/packages/abstract-client/src/error.rs
@@ -28,4 +28,13 @@ pub enum AbstractClientError {
 
     #[error("Account is Renounced and does not have an owner.")]
     RenouncedAccount {},
+
+    #[error("Can't retrieve Account for unclaimed namespace \"{namespace}\".")]
+    NamespaceNotClaimed { namespace: String },
+
+    #[error("Can't add custom funds when using auto_fund.")]
+    FundsWithAutoFund {},
+
+    #[error("Account creation auto_fund assertion failed with required funds: {0:?}")]
+    AutoFundsAssertFailed(Vec<cosmwasm_std::Coin>),
 }