+++ integrations/astroport-adapter/src/staking.rs
@@ -57,7 +57,7 @@ impl CwStakingCommand for Astroport {
         &mut self,
         deps: Deps,
         _env: Env,
-        _info: Option<cosmwasm_std::MessageInfo>,
+        _addr_as_sender: Option<Addr>,
         ans_host: &AnsHost,
         _version_control_contract: VersionControlContract,
         lp_tokens: Vec<AssetEntry>,