+++ framework/CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   
 - Added a `.execute` method on the AuthZ API to execute `CosmosMsg` types on behalf of a granter.
 - Add IBC helpers to account client.
+- Abstract Client builder: register dexes on ANS
+- `.sub_accounts` method on `Account` for getting Abstract Client Sub Accounts
+- Publish adapter method of Abstract Client Publisher now returns Adapter object
+- Added a `.account_from` method on the `AbstractClient` for retrieving `Account`s.
+- Creating Sub Account from `AbstractClient` Account builder.
+- Installing apps and adapters for `AbstractClient` Account builder
+- Attaching funds to account creation on `AbstractClient` Account builder
 
 ### Changed
 
@@ -19,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+- Creating sub-account from account factory is restricted. Use Create Sub Account method of the manager instead
 
 ## [0.20.0] - 2024-01-24